# hw2/openapi/marketplace.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:8000

tags:
  - name: Auth
  - name: Products

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and get tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error_code: TOKEN_INVALID
                message: Invalid credentials
                details: {}

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Refresh token invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error_code: REFRESH_TOKEN_INVALID
                message: Refresh token is invalid
                details: {}

  /products:
    get:
      tags: [Products]
      summary: List products with pagination and filtering
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - in: query
          name: status
          required: false
          schema:
            $ref: "#/components/schemas/ProductStatus"
        - in: query
          name: category
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Page of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"

    post:
      tags: [Products]
      summary: "Create product (SELLER: only own, ADMIN: any)"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "200":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/TokenInvalid"
        "403":
          $ref: "#/components/responses/AccessDenied"

  /orders:
    post:
      security:
        - BearerAuth: [ ]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Business conflict
        '429':
          description: Rate limit exceeded
  /orders/{id}:
    put:
      security:
        - BearerAuth: [ ]
      summary: Update order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    get:
      security:
        - BearerAuth: [ ]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

  /orders/{id}/cancel:
    post:
      security:
        - BearerAuth: [ ]
      summary: Cancel order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'


  /promo-codes:
    post:
      security:
        - BearerAuth: [ ]
      summary: Create promo code (SELLER, ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreateRequest'
      responses:
        '200':
          description: Promo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'



  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/ProductNotFound"

    put:
      tags: [Products]
      summary: "Update product (SELLER: only own, ADMIN: any)"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/TokenInvalid"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/ProductNotFound"

    delete:
      tags: [Products]
      summary: Soft delete product (set status ARCHIVED)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Archived
        "401":
          $ref: "#/components/responses/TokenInvalid"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/ProductNotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error_code: VALIDATION_ERROR
            message: Validation failed
            details:
              fields:
                - field: name
                  message: must not be empty

    AccessDenied:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error_code: ACCESS_DENIED
            message: Access denied
            details: {}

    TokenInvalid:
      description: Token invalid or expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            expired:
              value:
                error_code: TOKEN_EXPIRED
                message: Access token expired
                details: {}
            invalid:
              value:
                error_code: TOKEN_INVALID
                message: Access token invalid
                details: {}

    ProductNotFound:
      description: Product not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error_code: PRODUCT_NOT_FOUND
            message: Product not found
            details: {}

  schemas:
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    UserRole:
      type: string
      enum: [USER, SELLER, ADMIN]

    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 6
          maxLength: 72
        role:
          $ref: "#/components/schemas/UserRole"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 6
          maxLength: 72

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 10

    AccessTokenResponse:
      type: object
      required: [access_token]
      properties:
        access_token:
          type: string

    TokenPairResponse:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    UserResponse:
      type: object
      required: [id, email, role, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]

    ProductCreate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: "#/components/schemas/ProductStatus"

    ProductUpdate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: "#/components/schemas/ProductStatus"

    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: decimal
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: "#/components/schemas/ProductStatus"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductPageResponse:
      type: object
      required: [items, totalElements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
    OrderStatus:
      type: string
      enum:
        - CREATED
        - PAYMENT_PENDING
        - PAID
        - SHIPPED
        - COMPLETED
        - CANCELED

    DiscountType:
      type: string
      enum:
        - PERCENTAGE
        - FIXED_AMOUNT

    UserOperationType:
      type: string
      enum:
        - CREATE_ORDER
        - UPDATE_ORDER

    OrderItemRequest:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
          maximum: 999

    OrderCreateRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true

    OrderUpdateRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
        price_at_order:
          type: number
          format: decimal

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/OrderStatus'
        promo_code_id:
          type: integer
          format: int64
          nullable: true
        total_amount:
          type: number
          format: decimal
        discount_amount:
          type: number
          format: decimal
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PromoCodeCreateRequest:
      type: object
      required:
        - code
        - discount_type
        - discount_value
        - min_order_amount
        - max_uses
        - valid_from
        - valid_until
      properties:
        code:
          type: string
          minLength: 4
          maxLength: 20
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
          minimum: 0.01
        min_order_amount:
          type: number
          format: decimal
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time

    PromoCodeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: decimal
        min_order_amount:
          type: number
          format: decimal
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean